requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.23.0
confection==1.3.3
cymem==2.0.13
dj-database-url==3.1.2
Django==6.0.4
django-debug-toolbar==6.3.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.4.4
packaging==26.2
preshed==3.0.13
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
python-decouple==3.8
requests==2.33.1
rich==15.0.0
setuptools==82.0.1
shellingham==1.5.4
smart_open==7.6.0
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.3
thinc==8.3.13
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==1.0.0
whitenoise==6.12.0
wrapt==2.1.2
